requirements.txt
Click==7.0
docutils==0.15.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
lockfile==0.12.2
MarkupSafe==1.1.1
python-daemon==2.2.4
PyYAML==5.3
Werkzeug==0.16.0
